analytics_llm_Gemini/requirements_gemini.txt
google-generativeai>=0.7.0
duckdb>=0.10.0
streamlit>=1.32.0
pandas>=2.0.0
tabulate>=0.9.0
